Changes in JDBCStorageDelegate:
Removed CachedStatement from the methods (readAllExpressionFor and
readExpressionFor) so that new PreparedStatement is return every time.

Reasons:
According to the Postgres JDBC Documentation, a ResultSet is
automatically closed When the Statement used to create it, send another
query.
Postgres JDBC Documentation Link
https://jdbc.postgresql.org/documentation/head/resultset.html
So in a scenario where multiple processes are calling the same store,
they will use the same PreparedStatement.
The problem is when one process is still using his ResultSet while the
other process sends his query.
Removing the CachedStatement from the method guarantees that every
process will get a new PreparedStatement, therefore it will not
interrupt the ResultSet of another.